Scale using RC instead of a DC
We should really modify a Replication Controller instead of a Deployment
Config in order to change the number of running replicas.

As the name of a RC is derived from DC (for instance if DC name is "app",
RC name would be "app-1"), we cannot rely on the raw DC name. Instead, we
use DC name as a prefix to find the RC object we want to modify.

Also, the Checker function must count the exact number of replicas. Fix
this as well.