Add a "labels" parameter in OpenShiftHandle.execute(), making it more generic
Currently execute() works only in the same deployment as the test. We have
now another use case that is the test running in a pod and the images we
are testing living in other pods.

So, by passing the labels to it we leave to callers the task of finding the
pods they want to "execute" stuff in.