Add a labels parameter to adapter.getLog(), to better match a pod
The prefix argument, by itself is poor to find a pod, because you can
have many pods starting with the same prefix. For instance, in a
deployment with EAP+AMQ you can have 3 pods named:
 - eap-app-1-c7j8j
 - eap-app-amq-1-xx8xz
 - eap-app-1-build

All of them starting with "eap-app".

A "labels" argument then is added to better handle this situation. If
passed it will be used along with the prefix to find the pod.