Drop an unsed argument in deleteEnvironment()
That argument sometimes - for example if there's no template
to instantiate - can not be injected. As such, it would receive
a value of null.

Worse, it will block the method from being executed, because the
arquillian behavior for methods with arguments that cannot be
injected is simply not to invoke them.

If this method (deleteEnvironment) is not invoked, then the cleanup
will not be done, leaving some resources behind.

So, if there are other tests to be executed after this and if they
have resources that clash with the previous one, they will fail due
to an attempt to create a duplicated resource.

Closes #33.