We should scale DC, not RC
It's documented here: https://docs.openshift.org/latest/dev_guide/deployments/how_deployments_work.html#what-is-a-deployment
In short: We shouldn't touch RC's that are managed by a DC.

In order to be able to scale DC, we need to bump the
kubernetes-[api|client] library versions. As a side effect
of this bump we needed to change the way RoleBindings are
created.