Don't check for Persistent Volumes beforehand
This requires special privileges on the cluster, which is not
a good thing to depend on.

Regular users are able to use PV's through PVC's, but are not
able to list all PV's, which is what this feature does.

Being able to use the PV's, thus allowing the user to run the
tests are enough for our case use.

We can assume - and document - that the user is responsible
for configuring the Persistent Volumes on the cluster before
running the tests. This is out of the scope for the testsuite (
at least for now).