Initialize 'templates' with an empty list instead of null
So that it can be safely iterated anywhere in the code.

Particularly, this fixes a NPE when the deploy fails in the
method createEnvironment() and then deleteEnvironment() gets
called. It iterates over the templates field, which has the
null value, once the method initializes it was never called
because the deployment failed.