Override getToken() to return the client's token
If we have not explicitly provide a token, we still can get
the one being used by the client, if it exists.

So, from a library's user perspective, getToken() only returns
an empty string if 1) the user hasn't provided one and 2) the
openshift client could not retrieve one from the user's config
file (usually ~/.kube/config).